Add img shortcode
The short code adds an image that takes up the full width of the screen on mobile and a third of the width on desktop.  The image has a grey border that's highlighted on mouseover.

The home page also now shows the date of the the most recent post.  The title of the most recent post is no longer a link, because it was confusing.